extract response shared context
