+++ spec/spec_helper.rb
@@ -1,6 +1,7 @@
 require 'iban_calculator'
 
 require_relative 'shared/payload'
+require_relative 'shared/response'
 
 RSpec.configure do |config|
   config.filter_run :focus
